Edited borrower_server, settings, lendo_server and pdf_server
